+++ client/src/App/index.js
@@ -4,9 +4,8 @@ import {BrowserRouter as Router, Route } from 'react-router-dom';
 import Intro from './views/intro';
 import Landing from './views/landing';
 import Fiche from './views/fiche';
-import ChartStats from './components/ChartStats/ChartStats';
-import Chart from './views/graph/Chart';
-import SearchBar from './views/SearchBar/SearchBar';
+import Chart from './views/graph';
+import SearchBar from './views/SearchBar';
 import Integration from './views/integration';
 
 import './styles/main.scss';
@@ -16,10 +15,9 @@ class App extends Component {
         return(
             <Router>
                 <>
-                    <Route exact path="/" component={Intro} /> {/* root page */}
-                    <Route path="/Landing" component={Landing} />
+                    <Route path="/Intro" component={Intro} /> {/* root page */}
+                    <Route exact path="/" component={Landing} />
                     <Route path="/Fiche" component={Fiche} />
-                    <Route path="/ChartStats" component={ChartStats} />
                     <Route path="/Chart" component={Chart} />
                     <Route path="/SearchBar" component={SearchBar} />
                     <Route path="/Integration" component={Integration} />
