+++ client/src/App/scripts/CloudOfChamp.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+// import * as d3 from "d3";
+
+class CloudChart extends Component {
+
+    render() {
+        return (
+            <>
+
+            </>
+        )
+    }
+}
+
+export default CloudChart;
