+++ client/src/App/views/SearchBar/index.jsx
