+++ client/src/App/views/intro/intro.scss
@@ -36,18 +36,21 @@
     }
 }
 
-.img-left {
-    position: absolute;
-    height: 662px;
-    top: 50%;
-    left: 0;
-    transform: translateY(-50%);
-}
-
+.img-left,
 .img-right {
     position: absolute;
     height: 662px;
     top: 50%;
-    right: 0;
     transform: translateY(-50%);
+    img {
+        height: 100%;
+    }
+}
+
+.img-left {
+    left: 0;
+}
+
+.img-right {
+    right: 0;
 }
