+++ client/src/App/views/landing/index.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import CloudChart from '../../scripts/CloudOfChamp';
 
 import Header from '../../components/header';
 import Filter from '../../components/filter';
@@ -6,12 +7,18 @@ import Modal from '../../components/infoModal';
 
 import './landing.scss';
 
+
 class Landing extends Component {
 
+
     render() {
+
         return (
             <>
                 <Header />
+
+                <CloudChart />
+
                 <Filter />
                 <Modal />
             </>
